+++ tool/travis.sh
@@ -31,5 +31,6 @@ if [ "$COVERALLS_TOKEN" ] && [ "$TRAVIS_DART_VERSION" = "dev" ]; then
     --in=var/coverage.json \
     --out=var/lcov.info \
     --packages=.packages \
-    --report-on=lib
+    --report-on=lib \
+    --check-ignore
 fi